Don’t gossip graft induced broadcasts
When a peer receives a IHAVE message that it doesn’t
know about it will request a GRAFT from the originator.
When the BROADCAST that that GRAFT induced arrives
it isn’t desired for the peer to propagate it as if
it was a regular broadcast. The rationale being that this
was a pull operation made by the peer that only concerns
him, if the gossip is made throughout the overlay it might
generate unnecessary PRUNE messages from peers that
had already got the message before from the GRAFT
originator.