Drop stale outstanding lazy pushes
Over the course of an execution a peer accumulates several
lazy pushes to be sent on the lazy tick. Meanwhile it will receive
lazy pushes from other peers, these messages might make some
of the lazy push messages irrelevant since they already subsume
the outstanding ones. The intuition here is that if someone tells me
something it is irrelevant for me to tell that someone a subset of that
thing.

Expose a new callback method that allows comparing two message ids
and return a boolean to inform wether the first subsumes the second
(ie. the first is causally newer than the second).