Have broadcast_data callback run on the worker
Instead of having run on the client and then
sending out the message id and payload, send
the entire message to the broadcast worker and
have it invoke the callback method there.

Having it separate (as was before) and if the
callback method needed to write to some storage
would present a race condition since the
`broadcast_data` and `merge` callback methods would
run on different processes.