Marshal/Unmarshal broadcast data support
Marshaling/Unmarshaling of broadcast data is necessary to allow
for bulk handling. The application that intends to broadcast data in
bulk will need to define these methods that convert that into opaque
data and back into a list of messages to be handled individually.
This is basically a pack/unpack procedure before sending down
the wire.